requirements.txt
# ── Core Web App ──
streamlit>=1.32,<2.0
plotly>=5.18,<6.0

# ── Data & Finance ──
yfinance>=0.2.36,<1.0
pandas>=2.1,<3.0
pandas-ta>=0.3.14b1
numpy>=1.24,<2.0

# ── ML / Deep Learning ──
tensorflow>=2.15,<3.0
scikit-learn>=1.3,<2.0
xgboost>=2.0,<3.0
joblib>=1.3

# ── NLP / Sentiment ──
transformers>=4.36,<5.0
torch>=2.1,<3.0

# ── Computer Vision ──
opencv-python-headless>=4.8,<5.0
matplotlib>=3.8,<4.0
Pillow>=10.0,<11.0

# ── APIs & Networking ──
google-generativeai>=0.4,<1.0
requests>=2.31,<3.0
praw>=7.7,<8.0

# ── Config ──
python-dotenv>=1.0,<2.0

# ── Data Validation (new) ──
pandera>=0.18,<1.0
